// apps/client/src/types/types.ts
export type ContainerSize = Database["public"]["Enums"]["container_size"];
export type UpdateSettingsInfo =
  Database["public"]["Tables"]["settings"]["Update"];

export type ProfileRow = ToCamelCase<SupaProfileRow>;

export interface OrderStatistics {
  orders: number;
  mealCount: number;
  veggieMeals: number;
  thankYouBags: number;
  totalProteinWeight: number;
  teriyakiCuppyCount: number;
  extraRoastedVeggies: number;
  proteinCubes: Record<string, number>;
  containers: Partial<Record<ContainerSize, number>>;
  proteins: IngredientAmounts;
  veggieCarbs: IngredientAmounts;
}
export interface Order {
  fullName: string;
  itemName: string;
  container: ContainerSize;
  weight: number;
  flavor: string;
  flavorLabel: string;
  protein: string;
  proteinLabel: string;
  quantity: number;
}
export interface OrderError {
  error: Error | null;
  message: string;
  order: Order;
}
export interface IngredientAmounts {
  [name: string]: {
    label: string;
    amount: number;
    lbsPer: number;
    units: string;
  };
}
export interface Meal {
  protein: string;
  proteinLabel: string;
  flavor: string;
  flavorLabel: string;
  originalWeight: number;
  weight: number;
  weightLbOz: string;
  backstockWeight: number;
  cookedWeight: number;
  displayColor: string | null;
}
export interface PullListDatas {
  label: string;
  sunday: string;
  monday: string;
}
// ...
export interface OrderReportInfo {
  orderData: Order[];
  stats: OrderStatistics;
  orderErrors: OrderError[];
  meals: Meal[];
  usedBackstockIds: Set<number>;
  pullListDatas: PullListDatas[];
}